+++ src/main/kotlin/minesweeper/domain/Position.kt
