+++ src/main/kotlin/minesweeper/domain/Positions.kt
