+++ src/main/kotlin/minesweeper/domain/RandomPositionGenerator.kt
@@ -2,7 +2,7 @@ package minesweeper.domain
 
 import kotlin.random.Random
 
-class RandomPositionGenerator: PositionGenerator {
+class RandomPositionGenerator : PositionGenerator {
 
     override fun generate(end: Int): Int {
         return Random.nextInt(MINIMUM_RANDOM_NUMBER, end)
