+++ src/main/kotlin/minesweeper/view/Reception.kt
