+++ src/main/kotlin/minesweeper/view/ResultView.kt
