+++ src/test/kotlin/minesweeper/MarkerTest.kt
@@ -11,7 +11,7 @@ class MarkerTest {
 
     @ParameterizedTest
     @CsvSource(value = ["-10, 2", "0, 5", "2, 0", "5, -10"])
-    fun `높이 및 너비가 자연수가 아닌경우 예외처리된다`(height : Int, vertical: Int) {
+    fun `높이 및 너비가 자연수가 아닌경우 예외처리된다`(height: Int, vertical: Int) {
         assertThatIllegalArgumentException().isThrownBy {
             Marker(height, vertical)
         }
@@ -24,8 +24,8 @@ class MarkerTest {
         val ground = marker.generateAllGround()
 
         assertThat(ground).containsExactly(
-                listOf("C ", "C "),
-                listOf("C ", "C ")
+            listOf("C ", "C "),
+            listOf("C ", "C ")
         )
     }
 }
