+++ src/test/kotlin/minesweeper/domain/PositionsTest.kt
